Rename result_store package to results_store
Renames the physical folder from result_store to results_store to match results commands, and updates all internal import references in the codebase and test files.

TAG=agy

CONV=0be99d0f-5d5b-4387-bfcf-d83b50088a37
