+++ llmdbenchmark/results_store/client/__init__.py
@@ -1,8 +1,8 @@
 """Storage clients package for result store."""
 
-from llmdbenchmark.result_store.client.base import StorageClient
-from llmdbenchmark.result_store.client.gcs import GCSClient
-from llmdbenchmark.result_store.client.gcs_proxy import GCSProxyClient
+from llmdbenchmark.results_store.client.base import StorageClient
+from llmdbenchmark.results_store.client.gcs import GCSClient
+from llmdbenchmark.results_store.client.gcs_proxy import GCSProxyClient
 
 
 def get_storage_client(uri: str) -> StorageClient:
