+++ llmdbenchmark/results_store/commands/rm.py
@@ -2,8 +2,8 @@
 
 import fnmatch
 from pathlib import Path
-from llmdbenchmark.result_store.workspace import WorkspaceManager
-from llmdbenchmark.result_store.commands import register_command
+from llmdbenchmark.results_store.workspace import WorkspaceManager
+from llmdbenchmark.results_store.commands import register_command
 
 
 @register_command("rm")
