+++ llmdbenchmark/results_store/commands/status.py
@@ -2,10 +2,10 @@
 
 import sys
 from pathlib import Path
-from llmdbenchmark.result_store.store import StoreManager, StoreNotFound
-from llmdbenchmark.result_store.workspace import WorkspaceManager
-from llmdbenchmark.result_store.utils import color_pad
-from llmdbenchmark.result_store.commands import register_command
+from llmdbenchmark.results_store.store import StoreManager, StoreNotFound
+from llmdbenchmark.results_store.workspace import WorkspaceManager
+from llmdbenchmark.results_store.utils import color_pad
+from llmdbenchmark.results_store.commands import register_command
 
 
 @register_command("status")
